+++ DatadogRUM/Sources/RUMMonitor/RUMCommand.swift
@@ -245,6 +245,26 @@ internal struct RUMAddCurrentViewAppHangCommand: RUMErrorCommand {
     let missedEventType: SessionEndedMetric.MissedEventType? = .error
 }
 
+internal struct RUMAddCurrentViewMemoryWarningCommand: RUMErrorCommand {
+    var time: Date
+    var attributes: [AttributeKey: AttributeValue]
+    let canStartBackgroundView = false
+    let isUserInteraction = false
+
+    let message: String
+    let type: String?
+    let stack: String?
+    let category: RUMErrorCategory = .memoryWarning
+    let isCrash: Bool? = false
+    let source: RUMInternalErrorSource = .source
+    let errorSourceType: RUMErrorEvent.Error.SourceType = .ios
+    let threads: [DDThread]?
+    let binaryImages: [BinaryImage]?
+    let isStackTraceTruncated: Bool?
+
+    let missedEventType: SessionEndedMetric.MissedEventType? = .error
+}
+
 internal struct RUMAddViewTimingCommand: RUMCommand, RUMViewScopePropagatableAttributes {
     var time: Date
     var attributes: [AttributeKey: AttributeValue]